drf_jwt/api/views/test_views/test_views.py
# Create your views here.
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework import status
from rest_framework.permissions import AllowAny, IsAuthenticated, IsAdminUser, IsAuthenticatedOrReadOnly
from config.task import generate_song

class TestView(APIView):
    permission_classes = [IsAuthenticated]
    def post(self, request, *args, **kwargs):
        email = request.data['email']
        task = generate_song.delay(email=email)
        

        return Response(status=status.HTTP_200_OK)