drf_jwt/requirements.txt
mysqlclient==2.0.3
Django==3.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pyyaml==6.0
celery==5.2.7
flower==1.2.0